% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altTreat.R
\name{altTreat}
\alias{altTreat}
\title{Checking for alternate treatments using random forests.}
\usage{
altTreat(
  data,
  treatment,
  controls,
  ntree = 100,
  zthresh = 1,
  out = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{data}{is the universe of all variables considered, which may include
both controls and excluded variables (e.g. variables eliminated by a model
selection algorithm).}

\item{treatment}{is the name of the treatment variable contained in
\code{data}.}

\item{controls}{is a list() of names of selected controls contained in
\code{data}.}

\item{ntree}{is the numerical number of trees to use in the random forest
models.}

\item{zthresh}{is the z-score used as a threshold for flagging a variable
for suspicion as an alternative treatment, feedback, or intermediate bias.}

\item{out}{is a logical indicating whether to create a file named
"alternateTreatments.csv" in the current working directory that contains
the variables flagged as alternate treatments.}

\item{parallel}{is a logical indicating whether to use parallel processing}
}
\value{
\code{altTreat} returns a list that contains the following components:
\item{treatment}{name of the treatment variable.}
\item{controls}{the list of control variables.}
\item{excludedVariables}{the list of variables in data not included as
controls.}
\item{alternateTreatments}{the list of possible alternate treatments.}
\item{impZ}{the scaled variable importance matrix between the excluded
variables and the treatment,}
\item{zThresh}{the z-score threshold for suspicion.}
}
\description{
This function checks whether other variables that a previous process (e.g.
researcher "intuition", a model-selection algorithm) excluded might in
fact be a more useful treatment. In essence, it checks whether the current
treatment acted as a "collider" between eliminated variables and the
outcome (and hence may have influenced its elimination).
}
