% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eics.r
\name{eics}
\alias{eics}
\title{Empirically-Informed Covariate Selection (EICS) using random forests.}
\usage{
eics(
  data,
  outcome,
  treatment,
  controls,
  ntree = 100,
  zthresh = 1,
  out = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{data}{is the universe of all variables considered, which may include
both controls and excluded variables (e.g. variables eliminated by a model
selection algorithm).}

\item{outcome}{is the name of the outcome variable contained in \code{data}.}

\item{treatment}{is the name of the treatment variable contained in
\code{data}.}

\item{controls}{is a list() of names of selected controls contained in
\code{data}.}

\item{ntree}{is the numerical number of trees to use in the random forest
models.}

\item{zthresh}{is the z-score used as a threshold for flagging a variable
for suspicion as an alternative treatment, feedback, or intermediate bias.}

\item{out}{is a logical indicating whether to create a file named
"alternateTreatments.csv" in the current working directory that contains
the variables flagged as alternate treatments.}

\item{parallel}{is a logical indicating whether to use parallel processing}
}
\value{
\code{eics} returns a list that contains the following components:
\item{treatment}{name of the treatment variable.}
\item{controls}{the list of control variables.}
\item{excludedVariables}{the list of variables in data not included as
controls.}
\item{alternateTreatments}{the list of possible alternate treatments.}
\item{feedbackVariables}{the list of possible feedback variables.}
\item{intermediateVariables}{the list of possible intermediate variables.}
\item{alternateOutcomes}{the list of possible alternate outcomes.}
\item{zThresh}{the z-score threshold for suspicion.}
\item{impZ}{the scaled variable importance matrix between the excluded
variables and the treatment,}
\item{zMatrix}{the scaled variable importance matrix between the treatment,
controls, and outcome.}
}
\description{
This function checks all three possible violations of the backdoor
criterion by:
\enumerate{
\item Checking whether an eliminated variable might be superior to the
current treatment (separate function: altTreat);
\item Checking whether one or more control variables selected for the model
create a feedback loop with the current treatment;
\item Checking whether one or more control variables selected create intermediate
bias with the current treatment.
}
}
